perf(explore/programs): convert page to RSC, pre-warm React Query with server data
`app/explore/programs/page.tsx` was a 514-line `"use client"` page —
unlike the experts listing which is a proper RSC. Every navigation to
`/explore/programs` downloaded the full client bundle, mounted React,
fired multiple `useQuery` cold starts for trending / newest / topics /
stats, and only THEN painted anything. This is the dominant cause of
the perceived navigation lag the user reported.

Mirror the experts architecture:

- New thin async RSC `app/explore/programs/page.tsx` (~50 lines) calls
  `getCuratedPrograms("all", "trending", 8)`, `getCuratedPrograms("all",
  "newest", 8)`, `getTopicsWithCount("all")`, plus a small inline
  `prisma.classPlan.count() + prisma.webinarPlan.count()` for the hero
  stats — all in parallel via `Promise.all`. The data accessors already
  existed in `lib/data/explore-programs.ts`; they were just unused.
- New `app/explore/programs/ProgramsInteractiveContent.tsx` is the old
  `ProgramsContent` body, accepting `initialTrending` / `initialNewest`
  / `initialTopics` / `initialStats` as props. The hero stats no
  longer need the client `useEffect` fetch to `/api/programs/stats`.
- `useCuratedPrograms` and `useTopicsWithCount` get a new optional
  `initialData` parameter so the React Query cache is pre-warmed for
  the default `programType === "all"` query keys. Tab switches still
  trigger normal client fetches (the existing query-key plumbing stays
  unchanged).
- Tab-change URL sync uses `window.history.replaceState` instead of
  `router.replace` — same fix as the experts listing got in this PR
  (avoids the Next 15 `useSearchParams` reactivity that caused
  scroll-to-top jumps mid-typing).

Net effect on `/explore/programs`:
- First paint paints with real curated data instead of empty
  carousels + spinners
- Hero stats render with real numbers immediately, no fetch flash
- Bundle download / hydration still happens, but the visible content
  is already there
- Combined with the new top-level navigation progress bar, navigation
  to `/explore/programs` should feel instant rather than delayed
